requirements.txt
arduino-python3==0.6
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.2
decorator==5.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.5
django-rest-framework==0.1.0
django-stdimage==6.0.1
django-tables2==2.4.1
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
gunicorn==20.1.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyFirmata2==2.4.2
Pygments==2.12.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.2.0
requests==2.28.1
schedule==1.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
static3==0.7.0
tornado==6.2
traitlets==5.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.25.7
wcwidth==0.2.5
whitenoise==5.0.1
